+++ client/src/components/loginElement/LoginForm.tsx
@@ -1,7 +1,5 @@
 import React, { useState, useEffect } from "react";
 import { useForm } from "react-hook-form";
-import axios from "axios";
-import { useNavigate } from "react-router";
 import { Link } from "react-router-dom";
 import Oauth from "./Oauth";
 import PasswordChangeForm from "../myPage/infoChangeComponent/PasswordChange";
@@ -14,54 +12,16 @@ import {
   LoginInput,
   LogoImg,
 } from "./LoginForm.style";
+import { LoginApiCall } from "../../services/loginService";
 
 const LoginForm: React.FC = () => {
-  const [loginInfo, setLoginInfo] = useState<boolean>(false); //아아디,비밀번호가 입력돼었으면 true
-  const [id, setId] = useState<string>("");
-  const [pw, setPw] = useState<string>("");
   const [lostPw, setLostPw] = useState<boolean>(false); //비밀번호 변경 모달
 
-  const navigate = useNavigate();
   const { register, handleSubmit } = useForm();
-
-  useEffect(() => {
-    if (id && pw) {
-      setLoginInfo(true);
-    } else {
-      setLoginInfo(false);
-    }
-  }, [id, pw]);
-
-  const handleId = (value: string) => {
-    const idValue = value;
-    setId(idValue);
-  };
-  const handlePw = (value: string) => {
-    const pwValue = value;
-    setPw(pwValue);
-  };
   const handleModal = () => {
     setLostPw(!lostPw);
   };
-  const TestApiCall = async (data: object) => {
-    try {
-      const headers = {
-        "ngrok-skip-browser-warning": "1",
-      };
-      const response = await axios.post(
-        "https://0709-116-125-236-74.ngrok-free.app/auth/login",
-        data,
-        { headers: headers },
-      );
-      if (response.status === 200) {
-        console.log(console.log("성공"));
-        navigate("/feeds");
-        // response.data.access_token
-      }
-    } catch (err) {
-      console.log(err);
-    }
-  };
+
   return (
     <>
       <InputContainer>
@@ -69,34 +29,24 @@ const LoginForm: React.FC = () => {
         <form
           onSubmit={handleSubmit((data) => {
             console.log(data);
-            TestApiCall(data);
+            // LoginApiCall(data);
           })}
         >
           <LoginInput
             type="text"
             placeholder="이메일을 입력하세요."
-            {...register("email")}
-            onChange={(e) => {
-              handleId(e.target.value);
-            }}
+            {...register("email", { required: true })}
           />
           <LoginInput
             type="password"
             placeholder="비밀번호를 입력하세요."
-            {...register("password")}
-            onChange={(e) => {
-              handlePw(e.target.value);
-            }}
+            {...register("password", { required: true })}
           />
           <LostPassword onClick={handleModal}>
             비밀번호를 잊으셨나요?
           </LostPassword>
           {lostPw && <PasswordChangeForm setLostPw={setLostPw} />}
-          {loginInfo ? (
-            <LoginButtonOn type="submit">로그인</LoginButtonOn>
-          ) : (
-            <LoginButtonOff>로그인</LoginButtonOff>
-          )}
+          <LoginButtonOn type="submit">로그인</LoginButtonOn>
         </form>
         <div>
           <Oauth />
