+++ client/src/components/signUpElement/SignUpInputs.style.tsx
@@ -10,7 +10,7 @@ export const InputContainer = styled.div`
   }
   form {
     width: 100%;
-    height: 250px;
+    height: 230px;
     display: flex;
     flex-wrap: wrap;
     flex-direction: column;
