+++ client/src/components/signUpElement/SignUpInputs.tsx
@@ -1,7 +1,5 @@
 import React, { useState, useEffect } from "react";
 import { useForm } from "react-hook-form";
-import axios from "axios";
-import { useNavigate } from "react-router";
 import { ReactComponent as Message } from "../../assets/images/icons/signUpIcons/Message.svg";
 import { ReactComponent as Person } from "../../assets/images/icons/signUpIcons/Person.svg";
 import { ReactComponent as Lock } from "../../assets/images/icons/signUpIcons/Lock.svg";
@@ -14,9 +12,14 @@ import {
   TextInput,
   SubmitButton,
 } from "./SignUpInputs.style";
+import {
+  SignApiCall,
+  getAuthentication,
+  sendAuthentication,
+  checkNickName,
+} from "../../services/signUpService";
 
 const SignUpInputs = () => {
-  const navigate = useNavigate();
   const {
     register,
     handleSubmit,
@@ -27,7 +30,6 @@ const SignUpInputs = () => {
 
   const [allSelected, setAllSelected] = useState(false);
   const [isAuthentication, setIsAuthentication] = useState(false); //인증코드상태
-  const [isPwConfirm, setIsPwConfirm] = useState(false); //비번확인상태
 
   //각각 input 태그 value 호출
   const email = watch("email", "");
@@ -61,94 +63,12 @@ const SignUpInputs = () => {
     setValue("agree2", isChecked);
   };
 
-  //회원가입 버튼 함수
-  const TestApiCall = async (data: object) => {
-    try {
-      const headers = {
-        "ngrok-skip-browser-warning": "1",
-      };
-      const response = await axios.post(
-        "https://0709-116-125-236-74.ngrok-free.app/member/signup",
-        data,
-        { headers: headers },
-      );
-      if (response.status === 201) {
-        console.log("성공");
-        navigate("/");
-        // response.data.access_token
-      }
-    } catch (err) {
-      console.log(err);
-    }
-  };
-
-  //인증코드 보내는 버튼 함수
-  const getAuthentication = async (email: string) => {
-    try {
-      const request = await axios({
-        method: "post",
-        url: `https://0709-116-125-236-74.ngrok-free.app/member/emails/send-code?email=${email}`,
-      });
-      if (request.status === 200) {
-        console.log("인증코드를 보냈습니다.");
-      }
-    } catch (error) {
-      console.log(error);
-    }
-  };
-
-  //인증하기 버튼 함수
-  const sendAuthentication = async (email: string, authentication: string) => {
-    try {
-      const data = {
-        email: email,
-        authentication: authentication,
-      };
-      const headers = {
-        "ngrok-skip-browser-warning": "1",
-      };
-      const response = await axios.post(
-        "https://0709-116-125-236-74.ngrok-free.app/member/emails/check",
-        data,
-        { headers: headers },
-      );
-      if (response.status === 200) {
-        console.log("성공");
-        setIsAuthentication(true);
-        // response.data.access_token
-      }
-    } catch (err) {
-      console.log(err);
-    }
-  };
-
-  //닉네임 중복확인 버튼 함수
-  const checkNickName = async (nickName: string) => {
-    try {
-      const headers = {
-        "ngrok-skip-browser-warning": "1",
-      };
-      const response = await axios.post(
-        `https://0709-116-125-236-74.ngrok-free.app/member/name?nick=${nickName}`,
-        { headers: headers },
-      );
-      if (response.status === 200) {
-        console.log("성공");
-        setIsPwConfirm(true);
-        // response.data.access_token
-      }
-    } catch (err) {
-      console.log(err);
-    }
-  };
-
   return (
     <InputContainer>
       <h2>
         회원가입을 위해
         <br /> 정보를 입력해 주세요.
       </h2>
-      {/* 인증번호,중복확인 버튼구현해야함 */}
       <form
         onSubmit={(e) => {
           e.preventDefault();
@@ -181,7 +101,11 @@ const SignUpInputs = () => {
               placeholder="인증번호를 입력해주세요."
               autoComplete="off"
             />
-            <button onClick={() => sendAuthentication(email, authentication)}>
+            <button
+              onClick={() =>
+                sendAuthentication(email, authentication, setIsAuthentication)
+              }
+            >
               인증하기
             </button>
           </TextInput>
@@ -268,8 +192,9 @@ const SignUpInputs = () => {
         <SubmitButton
           type="submit"
           onClick={handleSubmit((data) => {
-            console.log(data);
-            TestApiCall(data);
+            isAuthentication === false
+              ? alert("이메일 인증은 필수입니다.")
+              : SignApiCall(data);
           })}
         >
           가입하기
